# ai-readiness-frontend/lib/openapi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Readiness Assessment API",
    "description": "Comprehensive API for AI readiness assessment platform including survey management, analytics, LLM analysis, and user management",
    "version": "1.0.0",
    "contact": {
      "name": "AI Readiness Support",
      "email": "support@ai-readiness.com",
      "url": "https://ai-readiness.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://ai-readiness-frontend.vercel.app/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/analytics/dashboard": {
      "get": {
        "summary": "Get dashboard analytics",
        "description": "Retrieve comprehensive analytics data for dashboard display",
        "operationId": "getDashboardAnalytics",
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "organizationId",
            "in": "query",
            "description": "Organization ID to filter results",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date for data range",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date for data range",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "metrics",
            "in": "query",
            "description": "Comma-separated list of metrics to include",
            "schema": {
              "type": "string"
            },
            "example": "overview,jtbd_trends,voice_quality"
          },
          {
            "name": "period",
            "in": "query",
            "description": "Aggregation period",
            "schema": {
              "type": "string",
              "enum": ["daily", "weekly", "monthly"],
              "default": "weekly"
            }
          },
          {
            "name": "export",
            "in": "query",
            "description": "Export data instead of displaying",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Export format",
            "schema": {
              "type": "string",
              "enum": ["json", "csv"],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard analytics data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardAnalytics"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                },
                "example": "overview,{\"totalResponses\":1247,\"completionRate\":87.5}"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Advanced analytics query",
        "description": "Submit complex analytics queries with custom filters and parameters",
        "operationId": "advancedAnalyticsQuery",
        "tags": ["Analytics"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvancedAnalyticsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Advanced analytics results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardAnalytics"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/llm/analyze": {
      "post": {
        "summary": "Analyze response with LLM",
        "description": "Perform Jobs-to-be-Done (JTBD) analysis on survey responses using Large Language Models",
        "operationId": "analyzeLLMResponse",
        "tags": ["LLM Analysis"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMAnalysisRequest"
              },
              "example": {
                "responseId": "uuid-response-id",
                "responseText": "We struggle with data quality and need better integration tools",
                "questionText": "What are your main challenges with current systems?",
                "expectedForce": "pain_of_old",
                "questionContext": "AI readiness assessment",
                "organizationId": "uuid-org-id",
                "surveyId": "uuid-survey-id"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "LLM analysis completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - LLM access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "LLM service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "LLM service health check",
        "description": "Check the status and health of LLM analysis services",
        "operationId": "llmHealthCheck",
        "tags": ["LLM Analysis"],
        "responses": {
          "200": {
            "description": "LLM service status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMHealthResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Supabase authentication"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Session cookie authentication"
      }
    },
    "schemas": {
...
      "DashboardAnalytics": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "overview": {
                "$ref": "#/components/schemas/OverviewMetrics"
              },
              "jtbd_trends": {
                "$ref": "#/components/schemas/JTBDMetrics"
              },
              "voice_quality": {
                "$ref": "#/components/schemas/VoiceMetrics"
              },
              "user_engagement": {
                "$ref": "#/components/schemas/EngagementMetrics"
              },
              "real_time": {
                "$ref": "#/components/schemas/RealTimeMetrics"
              }
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "organizationId": {
                "type": "string",
                "nullable": true
              },
              "dateRange": {
                "type": "object",
                "properties": {
                  "start": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "end": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                }
              },
              "period": {
                "type": "string",
                "enum": ["daily", "weekly", "monthly"]
              },
              "metrics": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "cacheTtl": {
                "type": "number",
                "description": "Cache TTL in seconds"
              }
            }
          }
        }
      },
      "OverviewMetrics": {
        "type": "object",
        "properties": {
          "totalResponses": {
            "type": "number"
          },
          "completionRate": {
            "type": "number",
            "format": "float",
            "description": "Completion rate as percentage"
          },
          "averageTime": {
            "type": "number",
            "format": "float",
            "description": "Average completion time in minutes"
          },
          "activeUsers": {
            "type": "number"
          },
          "growth": {
            "type": "object",
            "properties": {
              "responses": {
                "type": "number",
                "format": "float",
                "description": "Growth percentage"
              },
              "completion": {
                "type": "number",
                "format": "float"
              },
              "users": {
                "type": "number",
                "format": "float"
              }
            }
          }
        }
      },
      "JTBDMetrics": {
        "type": "object",
        "properties": {
          "forces": {
            "type": "object",
            "properties": {
              "push": {
                "type": "number",
                "description": "Push force score"
              },
              "pull": {
                "type": "number",
                "description": "Pull force score"
              },
              "habit": {
                "type": "number",
                "description": "Habit force score"
              },
              "anxiety": {
                "type": "number",
                "description": "Anxiety force score"
              }
            }
          },
          "readinessScore": {
            "type": "number",
            "format": "float",
            "description": "Overall AI readiness score"
          },
          "departmentBreakdown": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          }
        }
      },
      "VoiceMetrics": {
        "type": "object",
        "properties": {
          "totalRecordings": {
            "type": "number"
          },
          "averageDuration": {
            "type": "number",
            "format": "float",
            "description": "Average duration in minutes"
          },
          "transcriptionAccuracy": {
            "type": "number",
            "format": "float",
            "description": "Transcription accuracy percentage"
          },
          "qualityDistribution": {
            "type": "object",
            "properties": {
              "excellent": {
                "type": "number"
              },
              "good": {
                "type": "number"
              },
              "fair": {
                "type": "number"
              }
            }
          }
        }
      },
      "EngagementMetrics": {
        "type": "object",
        "properties": {
          "dailyActive": {
            "type": "number"
          },
          "weeklyActive": {
            "type": "number"
          },
          "retentionRate": {
            "type": "number",
            "format": "float"
          },
          "featureUsage": {
            "type": "object",
            "properties": {
              "surveys": {
                "type": "number"
              },
              "voice": {
                "type": "number"
              },
              "analytics": {
                "type": "number"
              }
            }
          }
        }
      },
      "RealTimeMetrics": {
        "type": "object",
        "properties": {
          "currentUsers": {
            "type": "number"
          },
          "activeSurveys": {
            "type": "number"
          },
          "systemHealth": {
            "type": "number",
            "format": "float",
            "description": "System health percentage"
          },
          "responseRate": {
            "type": "number",
            "format": "float",
            "description": "Real-time response rate"
          }
        }
      },
      "AdvancedAnalyticsRequest": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "dateRange": {
            "type": "object",
            "properties": {
              "start": {
                "type": "string",
                "format": "date-time"
              },
              "end": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "metrics": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["overview", "jtbd_trends", "voice_quality", "user_engagement", "real_time"]
            },
            "default": ["overview"]
          },
          "period": {
            "type": "string",
            "enum": ["daily", "weekly", "monthly"],
            "default": "weekly"
          },
          "export": {
            "type": "boolean",
            "default": false
          },
          "format": {
            "type": "string",
            "enum": ["json", "csv"],
            "default": "json"
          }
        }
      },
      "LLMAnalysisRequest": {
        "type": "object",
        "properties": {
          "responseId": {
            "type": "string",
            "description": "Survey response ID"
          },
          "responseText": {
            "type": "string",
            "maxLength": 5000,
            "description": "Text to analyze"
          },
          "questionText": {
            "type": "string",
            "maxLength": 1000,
            "description": "Original question text"
          },
          "expectedForce": {
            "type": "string",
            "enum": ["pain_of_old", "pull_of_new", "anchors_to_old", "anxiety_of_new", "demographic"],
            "description": "Expected JTBD force type"
          },
          "questionContext": {
            "type": "string",
            "description": "Additional context for analysis"
          },
          "organizationId": {
            "type": "string",
            "description": "Organization ID"
          },
          "surveyId": {
            "type": "string",
            "description": "Survey ID"
          }
        },
        "required": ["responseId", "responseText", "questionText", "expectedForce"]
      },
      "LLMAnalysisResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "analysisId": {
            "type": "string",
            "description": "Stored analysis ID"
          },
          "result": {
            "type": "object",
            "properties": {
              "force": {
                "type": "string",
                "enum": ["pain_of_old", "pull_of_new", "anchors_to_old", "anxiety_of_new", "demographic"]
              },
              "intensity": {
                "type": "number",
                "format": "float",
                "minimum": 0,
                "maximum": 100,
                "description": "Force intensity score (0-100)"
              },
              "confidence": {
                "type": "number",
                "format": "float",
                "minimum": 0,
                "maximum": 100,
                "description": "Analysis confidence score (0-100)"
              },
              "themes": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Identified themes in the response"
              },
              "summary": {
                "type": "string",
                "description": "Analysis summary"
              },
              "reasoning": {
                "type": "string",
                "description": "Explanation of the analysis"
              }
            }
          },
          "context": {
            "type": "object",
            "properties": {
              "questionContext": {
                "type": "string"
              },
              "employeeRole": {
                "type": "string"
              },
              "employeeDepartment": {
                "type": "string"
              },
              "organizationName": {
                "type": "string"
              }
            }
          }
        }
      },
      "LLMHealthResponse": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string",
            "example": "LLM Analysis API"
          },
          "status": {
            "type": "string",
            "enum": ["healthy", "degraded", "unhealthy"]
          },
          "latency": {
            "type": "number",
            "description": "Response latency in milliseconds"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "apiKeys": {
            "type": "object",
            "properties": {
              "openai": {
                "type": "string",
                "enum": ["configured", "missing"]
              },
              "anthropic": {
                "type": "string",
                "enum": ["configured", "missing"]
              }
            }
          },
          "config": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string"
              },
              "model": {
                "type": "string"
              }
            }
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["error"]
      },
      "ValidationErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          {
            "type": "object",
            "properties": {
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ]
      }
    }
  },
...
}